pyproject.toml
[project]
name = "runeform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.0.0",
    "anthropic>=0.40.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
]
